requirements.txt
# OCR App Requirements
# Python 3.10+

# Core Web Framework
Flask==3.1.3
Werkzeug==3.1.6

# OCR Libraries
paddleocr==3.2.0
paddlepaddle-gpu==3.2.0
paddlex==3.2.1

# Image Processing
opencv-contrib-python==4.10.0.84
Pillow==12.1.1
pdf2image==1.17.0
pypdfium2==5.5.0
imagesize==1.4.1

# PDF Processing
shapely==2.1.2
pyclipper==1.4.0

